refactor: migrate multilogger concurrency test to workers_test.go and remove dedicated test file
